Partial commit, can parse and match trades/quotes, but only for one stock
The issue is that the trades file contains each stock only over the
course of one day, whereas the quotes file has multiple trades
spread out over several days. This requires having an index for
each day and symbol in the quotes file to match multiple trading
days to a single quotes file.