Updated the script to handle trades with missing quotes entries
The script before assumed that the quotes file contained entries for
each trades entry. However the previous script, which detected errors,
found that the quotes files were missing numerous symbols that should
have been present. If there is a trade in the trades file with no
corresponding entry in the quotes file, the trade will be created
with values of 0 for the NBBx calculations