Updated regression analysis to use PLM for panel regressions
As well symbols banned during the ban period are more accurately banned by
marking each day the actual symbol was banned with a 1, rather than the
entire period for all symbols, as was done previously. As well added support
to generate a dataset for all symbols, not just banned symbols, and to use
the actual data rather than the difference between the ban and match.